Refactor preprocessors into separate classes
Extracted MergeSameStyleRuns and RemoveProofErrors into their own packages. This improves code organization and encourages separation of concerns, making the codebase easier to maintain and extend in the future.
